Add defunct google drive download implementation for future reference
This implementation didn't work. I was able to share the google drive
takeout folder with the service account but when takeout creates a new
takeout file in the folder, it isn't shared with the service account
even though the encompasing folder is. I wasn't able to find a
workaround for this.

Instead I'm going to have to unfortunately find another way. My next
method is going to be having a sync between s3 and gdrive.