Download latest takeout file from google drive
This is a rewrite to download from google drive (as originally intended)
rather than from S3 and depend on an external sync. The direction of
this job is also going to change. Rather than just downloading the file
we are going to sync it to S3. The files are substantially large and
working with S3 will be easier. A separate job will download the takeout
information from S3.