Throw error if no takeout files are found
This isn't exactly an error because it's normal for there to be no
takeout files on drive. If there aren't then we just don't do anything.
However, it's causing the app to fall through to tar which blows up
which isn't desirable.

Ideally there would be some control flow here where the application
would not continue but also wouldn't exit with a failure status code
because we don't want to consider this a failure.

Until this sort of incorrect feedback that the job failed is a problem,
I'm going to leave it as this.