Docker image to assist in testing container orchestration
This job takes input in the format of:
<sleep duration in seconds>|<output>|<exit code>

Its purpose is to be used in testing. Primarily conjob testing. It's
somewhat a replacement for the echo and sleep jobs with the addition of
allowing the exit code to be specified.

Rather than creating an exit job I think it's best to finally create a
single job which can be used for testing purposes as that's what the
echo and sleep jobs have been primarily used for. Any further testing
support can be added to this job instead of creating a new one.