Add image for assisting in testing docker in docker
This image can be run in a container to verify that running a nested
container works in a given environment (i.e. with the sysbox-runtime set
up properly).

It also lists existing images and containers from the cache. This can be
used to verify that mounting a volume for the docker cache is working as
expected.