Don't return null conclusions for in progress github workflow runs
It seems that if the workflow run is in progress, the conclusion is
null. This wasn't a problem before, but I changed how I was consuming
this data and things broke. I also think it's a better idea to not have
this interface potentially exposing null.

Instead it will now return an empty string if the conclusion is null.