Use input creds if specified instead of secrets file for gh build status
This allows different creds to be specified as input to the job. Since
the dashboard this is used on is public we just want to show public
repos there, but if the user has GitHub creds which can see private
repos, then they can pass them as input and this will return the build
status of private repos as well.

This enables me to see my private repo build status on personal
channels.