Report status from all builds, not just the latest
Before we were just getting the status of the latest build. However, if
we have multiple workflows this doesn't give an accurate representation
of the repos build status.

Now, we look at the status of all checks from every branch and if any
are failing we consider the whole repo to be failing.

- Since we used flat() we also needed to fix some compilation errors by
upgrading the node version and including the tsconfig.